refactor: clean screens
