--- config.json
@@ -1,5 +1,0 @@
-{
-  "domain": "cozy.works",
-  "name": "testchristophe",
-  "slug": "template"
-}
