--- src/navigator/index.js
@@ -1,3 +1,0 @@
-import Navigator from './AppNavigator'
-
-export default Navigator
