+++ src/screens/connectors/LauncherView.js
