--- src/screens/home/FileListPage.js
@@ -1,30 +1,0 @@
-import React from 'react'
-import {
-  SafeAreaView,
-  ScrollView,
-  StatusBar,
-  StyleSheet,
-  View,
-} from 'react-native'
-import FileList from '../../ui/FileList'
-
-const FileListPage = () => {
-  return (
-    <View style={styles.container}>
-      <StatusBar barStyle="dark-content" />
-      <SafeAreaView>
-        <ScrollView>
-          <FileList />
-          <FileList />
-          <FileList />
-        </ScrollView>
-      </SafeAreaView>
-    </View>
-  )
-}
-
-const styles = StyleSheet.create({
-  container: {},
-})
-
-export default FileListPage
