--- src/screens/home/index.js
@@ -1,5 +1,0 @@
-import Home from './Home'
-import FileListPage from './FileListPage'
-import Details from './Details'
-
-export {Home, FileListPage, Details}
