--- src/screens/konnectors/index.js
@@ -1,3 +1,0 @@
-import Connectors from './Connectors'
-
-export default Connectors
