--- src/screens/settings/index.js
@@ -1,3 +1,0 @@
-import Settings from './Settings'
-
-export {Settings}
