--- src/utils.js
@@ -1,3 +1,0 @@
-import {Platform} from 'react-native'
-
-export const isWebApp = () => Platform.OS === 'web'
